% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sts_get_metadata.R
\name{sts_get_metadata}
\alias{sts_get_metadata}
\title{Prepare Data For Analyses}
\usage{
sts_get_metadata(x, output = "SpeciesObservations")
}
\arguments{
\item{x}{the SynTreeSys database}

\item{output}{character. Should the plot metadat be returned (i.e.
'Plots') or the the species table (i.e. 'SpeciesObservations')
with the plot metadata? Defaults to 'SpeciesObservations'.}
}
\value{
A data frame
}
\description{
Simple function to get species abundances and plot
metadata together for analyses
}
