% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sts_get_sads.R
\name{sts_get_sads}
\alias{sts_get_sads}
\title{Get the Species Abundances Distributions per Group}
\usage{
sts_get_sads(
  x,
  minN = 0,
  maxN = Inf,
  minCutOff = 0,
  group.name = NULL,
  spp.name = "organismNameCurated",
  remove.indet = TRUE,
  abund.metric = "countsMeasurement",
  strata.names = NULL
)
}
\arguments{
\item{x}{the SynTreeSys database}

\item{minN}{numerical. The minimum number of individuals per plot.
Defaults to zero.}

\item{maxN}{numerical. The maximum number of individuals per plot.
Defaults to Inf.}

\item{minCutOff}{numerical. The minimum size criteria cutoff of the
plot (in mm). Defaults to zero.}

\item{group.name}{character. The STS column name with the categorie
used for grouping abundances.}

\item{spp.name}{character. The STS column name with the
(morpho)species name. Defaults to "organismNameCurated".}

\item{remove.indet}{logical. Should morphospecies not identified at
the species level be removed? Defaults to TRUE}

\item{abund.metric}{character. The STS column name with the
abundance metric to be used to build the SAD. Defaults to
"countsMeasurement"}

\item{strata.names}{vector of characters. The names of the strata
the should be kept in the analyses.}
}
\description{
Get the Species Abundances Distributions per Group
}
