% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntreesysTools-package.R
\docType{package}
\name{syntreesysTools-package}
\alias{syntreesysTools}
\alias{syntreesysTools-package}
\title{syntreesysTools: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SynTreeSys/syntreesysTools}
  \item Report bugs at \url{https://github.com/SynTreeSys/syntreesysTools/issues}
}

}
\author{
\strong{Maintainer}: Renato Lima \email{raflima@usp.br} (\href{https://orcid.org/0000-0002-1048-0138}{ORCID}) [copyright holder]

}
\keyword{internal}
